Latest/ewas_input/requirements.txt
Django>=1.8,<2.0
mysqlclient
pyexcel-xlsx
pyexcel-xls
django-excel
gunicorn
mysql-connector-python